참여자 보기 Dialog 작업 수정
- roomId단위 Long으로 변경
- 내보내기 기능 연결

+++ app/src/main/java/com/example/mnt_android/viewmodel/JoinRoomViewModel.kt
@@ -102,7 +102,7 @@ class JoinRoomViewModel(application: Application) : AndroidViewModel(application
         else
         {
 
-            repository.attendRoom(roomNum.toInt(), sf.getString("kakao_token", "null"))
+            repository.attendRoom(roomNum.toLong(), sf.getString("kakao_token", "null"))
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe(Action {
